Update README.md
Human skin is the largest organ in the body. It weighs between six and nine pounds and has a two-square-yard surface area.The skin separates the inner parts of the body from the outside world. It protects against fungal infection, germs, allergies, and viruses, as well as regulating body temperature. Swelling, burning, redness, and itching might be caused by situations that frustrate, modify the texture of the skin, or harm the skin. Dermatitis, hives, and other skin disorders can be caused by allergies, irritants, hereditary structure, and specific diseases and immune system problems. Many skin illnesses, such as acne, alopecia, ringworm, and eczema, have an impact on your appearance. Many different types of cancers can be caused by the skin. Nowadays, skin cancer is very common.
Skin cancer is the fourth greatest cause of non-fatal disease burden in the world. This burden affects both high- and low-income countries, indicating that skin cancer prevention should be prioritized. In this study, an intelligent diagnosis strategy for multi-class skin cancer categorization is provided. Nowadays almost everyone suffer from skin cancer.Various forms of skin cancer affect millions of individuals throughout the world. In recent years, computer-assisted skin examinations and characterization of skin cancer have increased. Examinations carried out include detecting and modeling human skin cancer. However, it is difficult to synthesize and analyze human skin because of the complexity of wrinkles, color differences, presence of hair, and other mitigation features, plus variations in dermatological diseases that occur due to geographical factors such as temperature, humidity and other environmental factors. On the other hand, detecting skin cancer is very time consuming and expensive at the same time. This research will help human to detect skin cancer in early stage. This research has been presented as an intelligent expert to perform the multi-class classification of skin cancer using deep learning algorithms. Another thing is, in the field of medical science 2 this research plays an important role. It is observed that better results and precision can be achieved when the algorithm is trained with more number of images.